+++ tests/sender/send_data.py
@@ -1,11 +1,11 @@
 import unittest
 import socket
-import time
+from pathlib import Path
 from ssl import CERT_NONE
 
-import devo.sender.data
 from devo.sender import Sender, SenderConfigTCP, SenderConfigSSL, \
     DevoSenderException
+from devo.sender.data import open_file
 from devo.common import get_log
 
 try:
@@ -592,3 +592,11 @@ class TestSender(unittest.TestCase):
                     crypto.load_certificate(
                         crypto.FILETYPE_PEM, str(_ca)))
         self.assertEqual(chain_certs, fake_chain_cert)
+
+    def test_open_file(self):
+        with self.assertRaises(FileNotFoundError) as result:
+            open_file(Path('wrong_file'), mode='r', encoding='utf-8')
+        with self.assertRaises(FileNotFoundError) as result:
+            open_file('wrong_file', mode='r', encoding='utf-8')
+        with self.assertRaises(DevoSenderException) as result:
+            open_file(55, mode='r', encoding='utf-8')